+++ sdk/resourcemanager/deviceprovisioningservices/armdeviceprovisioningservices/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Release History
 
+## 2.0.0-beta.1 (2023-05-02)
+### Breaking Changes
+
+- Type of `ErrorDetails.Code` has been changed from `*string` to `*int32`
+- Field `EnableDataResidency` of struct `IotDpsPropertiesDescription` has been removed
+
+### Features Added
+
+- New enum type `ManagedServiceIdentityType` with values `ManagedServiceIdentityTypeNone`, `ManagedServiceIdentityTypeSystemAssigned`, `ManagedServiceIdentityTypeSystemAssignedUserAssigned`, `ManagedServiceIdentityTypeUserAssigned`
+- New function `*IotDpsResourceClient.Failover(context.Context, string, string, CustomerInitiatedFailoverInput, *IotDpsResourceClientFailoverOptions) (IotDpsResourceClientFailoverResponse, error)`
+- New struct `CustomerInitiatedFailoverInput`
+- New struct `IotDpsPropertiesDescriptionDpsFailoverDescription`
+- New struct `ManagedServiceIdentity`
+- New struct `UserAssignedIdentity`
+- New field `DpsFailoverDescription`, `EnableCustomerInitiatedFailover`, `PortalOperationsHostName` in struct `IotDpsPropertiesDescription`
+- New field `Identity`, `Resourcegroup`, `Subscriptionid` in struct `ProvisioningServiceDescription`
+- New field `Resourcegroup`, `Subscriptionid` in struct `Resource`
+
+
 ## 1.1.0 (2023-03-28)
 ### Features Added